Refactored from Recording. Kanbas part Left.
