+++ src/Kanbas/Courses/Account/Profile.tsx
@@ -27,19 +27,19 @@ export default function Profile() {
             <h1>Profile</h1>
             {profile && (
                 <div>
-                    <input className="wd-username" value={profile.username}
+                    <input className="wd-username form-control mb-2" value={profile.username}
                         onChange={(e) => setProfile({ ...profile, username: e.target.value })} />
-                    <input className="wd-password" value={profile.password}
+                    <input className="wd-password form-control mb-2" value={profile.password}
                         onChange={(e) => setProfile({ ...profile, password: e.target.value })} />
-                    <input className="wd-firstname" value={profile.firstName}
+                    <input className="wd-firstname form-control mb-2" value={profile.firstName}
                         onChange={(e) => setProfile({ ...profile, firstName: e.target.value })} />
-                    <input className="wd-lastname" value={profile.lastName}
+                    <input className="wd-lastname form-control mb-2" value={profile.lastName}
                         onChange={(e) => setProfile({ ...profile, lastName: e.target.value })} />
-                    <input className="wd-dob" value={profile.dob}
+                    <input className="wd-dob form-control mb-2" value={profile.dob}
                         onChange={(e) => setProfile({ ...profile, dob: e.target.value })} type="date" />
-                    <input className="wd-email" value={profile.email}
+                    <input className="wd-email form-control mb-2" value={profile.email}
                         onChange={(e) => setProfile({ ...profile, email: e.target.value })} />
-                    <select className="wd-role" onChange={(e) => setProfile({ ...profile, role: e.target.value })}>
+                    <select className="wd-role form-control mb-2" onChange={(e) => setProfile({ ...profile, role: e.target.value })}>
                         <option value="USER">User</option>            <option value="ADMIN">Admin</option>
                         <option value="FACULTY">Faculty</option>      <option value="STUDENT">Student</option>
                     </select>
