+++ src/Kanbas/index.tsx
@@ -8,8 +8,9 @@ import * as client from "./Courses/client";
 import { useEffect, useState } from "react";
 import store from "./store";
 import { Provider } from "react-redux";
-import Account from "./Account";
-import ProtectedRoute from "./ProtectedRoute";
+import Account from "./Courses/Account";
+import ProtectedRoute from "./Courses/Account/ProtectedRoute";
+import Session from "./Courses/Account/Session";
 
 export default function Kanbas() {
     const [courses, setCourses] = useState<any[]>([]);
@@ -52,25 +53,27 @@ export default function Kanbas() {
 
     return (
         <Provider store={store}>
-            <div id="wd-kanbas">
-                <KanbasNavigation />
-                <div className="wd-main-content-offset p-3">
-                    <Routes>
-                        <Route path="/" element={<Navigate to="Dashboard" />} />
-                        <Route path="/Account/*" element={<Account />} />
-                        <Route path="Dashboard" element={<ProtectedRoute><Dashboard
-                            courses={courses}
-                            course={course}
-                            setCourse={setCourse}
-                            addNewCourse={addNewCourse}
-                            deleteCourse={deleteCourse}
-                            updateCourse={updateCourse} /></ProtectedRoute>} />
-                        <Route path="Courses/:cid/*" element={<ProtectedRoute><Courses courses={courses} /></ProtectedRoute>} />
-                        <Route path="Calendar" element={<h1>Calendar</h1>} />
-                        <Route path="Inbox" element={<h1>Inbox</h1>} />
-                    </Routes>
+            <Session>
+                <div id="wd-kanbas">
+                    <KanbasNavigation />
+                    <div className="wd-main-content-offset p-3">
+                        <Routes>
+                            <Route path="/" element={<Navigate to="Dashboard" />} />
+                            <Route path="/Account/*" element={<Account />} />
+                            <Route path="Dashboard" element={<ProtectedRoute><Dashboard
+                                courses={courses}
+                                course={course}
+                                setCourse={setCourse}
+                                addNewCourse={addNewCourse}
+                                deleteCourse={deleteCourse}
+                                updateCourse={updateCourse} /></ProtectedRoute>} />
+                            <Route path="Courses/:cid/*" element={<ProtectedRoute><Courses courses={courses} /></ProtectedRoute>} />
+                            <Route path="Calendar" element={<h1>Calendar</h1>} />
+                            <Route path="Inbox" element={<h1>Inbox</h1>} />
+                        </Routes>
+                    </div>
                 </div>
-            </div>
+            </Session>
         </Provider>
     );
 }
