+++ src/Kanbas/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import modulesReducer from "./Courses/Modules/reducer";
 import assignmentsReducer from "./Courses/Assignments/reducer";
-import accountReducer from "./Account/reducer";
+import accountReducer from "./Courses/Account/reducer";
 
 const store = configureStore({
     reducer: {
